Add question dialog for url clashes on renaming in TagChooser
Before this commit, if the tag was renamed so that the url
will clash with some other tag in the database, the TagChooser
wouldn't notice it, only notice that the rename didn't succeed
(but it didn't show any error message).
This commit ensures that in case of a name/url clash, TagChooser
will ask the user whether they want to remove the existing tag
or cancel the operation.