Don't allow custom All and All Untagged tags
Before this commit, Krita would have undefined behaviour
(in one case, just create a new tag) when encountering
a situation when the user asks for "All" or "All Untagged"
called tag (either during a rename or a creation).
This commit ensures that the user will be informed
that it's not possible, and the operation won't proceed.