Android: Use App-specific external storage for resources and config
The main benefit of this is that for most OEMs this allows the resources
directory to be mounted and edited. This doesn't require any additional
permissions for the devices we target.

Android warns that the External Storage may not be available
sometimes (eg when SD card is removed), in those cases we use the
Internal Storage instead.