StoryboardDocker: Code cleanup, polish pass.
Reduced the amount of repeated compile-time conditionals. Opted to make
  a utility class for the desired behavior with preprocessor conditional
  for easier reading.

Merged a few similar methods into one method for better reuse and improved legibility.
  Kept old function in place for compatabilty but simply calls the other singular method.

Improved behavior of queuing frames for rendering to acommodate for
  layers that have no keyframes, or for time-spans that encompass multiple
  non-aligned storyboard entries.

Fixed spontaneous crash in storyboard docker unit test due to improper
  deletion order.

TLDR: Some parts of the codebase should be easier to read now, and also be
more responsive for the user in certain edge cases where layers are being modified
and need to update multiple "scenes" as a result of the changes.