Add getResourceIdFromVersionedFilename to KisResourceCacheDb
Before this commit, there was no easy way to learn which resource
the specific resource file belonged to; this commit adds
a way to get the resource id from the database if you give it
any resource file (can be one that belongs to no-longer-current
version).

Example:
there is a resource called "My Best Preset" which has three versions.
The versions filenames are: "My_Best_Preset.kpp",
"My_Best_Preset.00001.kpp", "My_Best_Preset.00002.kpp".
The current version is the last one.
The function will give the same resource id for all filenames
described above no matter which filename belongs to the current
version (as it would be if the programmer would use a function like
Locator->resource or KisResourceModel->resource()).