Resource Manager no longer fails to add an inactive tag.
Previous behavior caused the text field in the "add tag" menu to
appear as though it fails to add a tag. Now we activate a tag
when the user tries to manually add it using the text field,
with the assumption that the user wants to use the tag again.

BUG:440615