Add KisResourceOverwriteDialog in a common area
Before this commit, all places where the user is asked
whether they want to overwrite a resource or not
would have to have their own dialogs.
This commit adds a common function to call to get the same
dialog.
It also adds a convenient function to check if the resource
is in the resource folder.
Since it's not a definitive, generic function (it doesn't
check bundles and other storages, only the folder storage)
it's just a helper function, hence it's here, and not in
the ResourceLocator.

CCMAIL:kde-i18n-doc@kde.org