Fix mirroring of Quick Engine brushes
Before this commit, the mirroring of Quick Engine brushes
was incorrect (different from how other brushes behaved).
That was because of the -0.5 pixel fix to align the outline
with the result. It was added before the mirrored points
were calculated, resulting in a wrong points.
This commit ensures that the -0.5 fix is applied only when
painting, which fixes the mirroring.

CCBUG:427831