Fix the presets become dirty on Krita start
There are some cycling signal dependencies in the resource view
initialization steps. The trace of the bug was like that:

1) The document adds its own storage
2) The tag model updates all the views with filterChanged signal
3) KisPatternChooser recieves resourceSelected() signal from
   KisResourceItemChooser
4) KisPatternChooser notifies the preset that its pattern has changed

This patch breaks this chain by ensuring that KisResourceItemChooser
does not emit resourceSelected() if the resource hasn't actually changed.
This solution is not ideal. Ideally we should properly do the cold-
initialization on connection of the view to the chooser and make sure
that the signals are emitted only when there is a real change in selection.
Right now this ideal solution doesn't work because there is no
cold-inialization at all, instead initialization relies on the signals.
Therefore, if I add deduplication of signals into
KisResourceItemListView::selectionChanged, the widgets are not initialized
properly.

CCBUG:440816