Brush Editor: Fix unchecking share curve resetting curve
Before this commit, from the user's perspective unchecking "share curve"
would result in the curves for all the sensors, including the current
sensor resetting to what they were before "share curve" was checked*.

This commit makes it so that the current curve and any active curves are
not reset (by setting them to the previously shared curve). Other curves
will still reset as before.

* If a new brush is saved or krita is restarted, disabled sensors will
  get reset to the default curve.