Don't search for occurrence of empty strings in resourceExists()
For filename there exists an automatic index due to the table condition:
UNIQUE("storage_id","resource_type_id","filename")
so the search is fast, and should solve Bug 440951.

But the search for md5 is about 2 orders of magnitude on a
resource collection with a few additional packs.
It's not clear yet why searching for name isn't nearly as slow,
despite also lacking an (auto-)index, according to the
"EXPLAIN QUERY PLAN" output.

BUG:440951