Fix typo in WindowLayoutResource sort screens lambda
Before this commit, if the screens had different X values,
the return value would be always false because the lambda
checked aRect.y() < aRect.y() which can never be true since
it's the same value on both sides.

This commit fixes it by replacing one of them with bRect.y().
The direction is consistent with the previous line.

BUG:421082