Fix a segfault in gamut mask toolbar
After deleting a gamut mask (or cancelling the creation of
a new gamut mask), using the gamut mask toggle button
in the gamut mask toolbar (available in circular artistic
color selectors) causes a segfault.

This is because the toolbar attempts to dereference a pointer to the
deleted gamut mask. This pointer should be nulled when the mask is
unset.

BUG:441122