Fix the usage of texture buffers in KisTextureTile
In the previous implementation we recreated the buffer on every upload,
which could be rather slow. The patch fixes it by allocating the buffer
only once (with max possible size) and reusing it every time.

WARNING: This patch expects the image to be RGB8! The hardcoded
pixel value must be fixed before the merging.