Fix default filename for Log Viewer not saving on Windows
Before this commit, if the user tried to save the log from the
Log Viewer on Windows, and just clicked "OK" without changing the
default filename, there would be an error of not being able to
save the file.
The reason for that was a colon in the filename, which is not
allowed on Windows.
This commit removes the colon by providing a custom format
for the date that is attached to the log filename.

BUG:440295