Limit frequency range based on resolution on screentone generator
Now the range of possible frequencies is limited to prevent the
creation of very large or very small cells.

The resolution is used to compute the range in a way that the
minimum frequency gives cells of 1000px and the maximum frequency
cells of 1px. [1, 1000] px is the range that can be set directly
in the size sliders (also in previous versions).

Before this patch a very low frequency could be set having as a
result very large cells (thousands of pixels). This produced crashes
when the template based equalization was used due to running out of
memory because of the allocation of very large template.