Indicate when H264 support is disabled in FFMpeg
A common mistake made by users is downloading or using an FFMpeg
version that has H264 disabled, then trying to create an MP4/MKV
only to get an "unknown error". Until the render UI is overhauled,
this is an interim solution to make it clearer for users that their
FFMpeg does not support MP4/MKV by putting "(MP4/MKV UNSUPPORTED)"
next to the version.

Also makes it more clear when FFMpeg is not found.