Fix touch pan gesture stopping when moving quickly
The code before this change ignores large movements to prevent canvas
from jumping at the end of a 3-finger gesture, which also blocks fast
panning. This removes the movement delta check in favour of checking the
number of touch points, so that fast movements will no longer interrupts
the pan gesture.

BUG: 441706