Implement KisAlgebra2D::cropLineToConvexPolygon
Before this commit, there was only cropLineToRect
which wasn't enough for some assistants which had limited area
of preview (which becomes a rotated rect if the user rotates the canvas,
which means it needs a custom function to cut the lines,
especially since we want to cut the lines to viewport as well.

This uses Cyrus-Beck algorithm.
https://en.wikipedia.org/wiki/Cyrus%E2%80%93Beck_algorithm