Fix precision issues with KisAlgebra2D::intersectLines
Sometimes on the edges of the "bounded line" in the intersectLines
the function wouldn't notice it because of some little precision
issues. This commit ensures that if the intersection point is
not further than an epsilon from the boundary point of the
bounded line, then it still will be considered
even if it is technically "outside" of the boundary.