Fix previews of various local assistants
Before this commit, some assistants wouldn't handle a rotated canvas
correctly (the preview would bleed out of the local area), and the
previews behaviour was inconsistent between different assistants.
This commit fixes the previews and makes it consistent:
- both the static (depending only on the assistant) and dynamic
(depending on the mouse position) previews are always visible
- both previews are always limited to the local area
(except for the parallel ruler which doesn't have any static preview
except for the exact line that was drawn, which is fully visible
no matter the local area - it could be fixed at a later date
if it turns out to be a big issue).

BUG:441212