Check that the model index is beneath colorchannel count.
It seems the real bug is that the thumbnails/model can take a bit to
update when generating the thumbnails, while the model/view is faster.
In this time, we shouldn't try to access information that doesn't exist.

BUG:442117


(cherry picked from commit 1eaf9dc7f624ef29d6358c243dd25437adb890b1)