Bugfix: Dockers cannot be attached when mainwindow is small
This is a workaround for the bug with docking system. When the onion
skins docker is enabled, the curves docker is shrunk below its minimum
size making it impossible for other dockers to attach. This seems to be
caused by the title bar of the docker not shrinking beyond the
minimumSizeHint().

BUG:441761