Fix imprecise dimensions when using image size dialog
Due to the way multiple units are handled, the result can be a tiny bit
smaller then the true value of e.g. 50% of 2816 px, despite qreal precision.
Furthermore, the spinbox will round to nearest integer when you
switch back to pixels, so the final value should consistently be rounded
instead of truncated for non-pixel input units.

BUG:442124