Implement a "better" stylus smoothing algorithm
The new algorithm just expects the tablet events to arrive at
a constant rate. We calculate average rate by using last 200 samples
(which corresponds to about 1-2 seconds) and then just use it to
calculate the speed between samples.

This algorithm bascially ignores the time values attached to the table
events themselves. They are not very reliable (somewhat okay on Linux
and completely unreliable on Windows).