OpenEXR: upgrade our Find module
Currently, we use Pixar's find module to manage OpenEXR 2, which cannot
find either v2.3 and lower or v3. The suggestion in the bug report was
to use ECM, which in turn has the following issues:

* It completely ignores the versioned suffix that OpenEXR applies
to its libraries; this implies that it's relying on Linux's symlinks to
point it to the correct library. This is a blocker for any level of
support for Windows.

* Secondly, but also less serious, it doesn't cover the IexMath and
IlmImfUtil subcomponents.

This commit replaces our Find module with ECM's, and makes the following
modifications.

1. Find and use any CMake config module that OpenEXR provides.
2. Failing that, construct the library suffix, and manually look up each component.
3. Add IexMath and IlmImfUtil to the list of libraries.
4. Add support for OpenEXR v3 and its list of components.

Additionally, I've cleaned up our find and linking logic.

BUG: 419277

CCMAIL: kimageshop@kde.org