Implement a GUI option for Brush Speed Smoothness
When we calculate the stylus speed we need to smooth values
a bit. Previuosly, Krita users expected rather huge amount of
smoothness, but now the default value is very low. If the user
want to make it larger, he/she is free to do that in the settings.
No restart is needed.

"Brush Speed Smoothness" value technically means how many hardware
samples we are going to average. It means that the value might have
different effect on different tablets, OSes and tablet APIs.

CC:kimageshop@kde.org