Fix frame cache invalidation with live preview on Windows.
I couldn't determine why this behavior is different between windows and
linux, but I was able to solve the issue. Basically, when we refresh
graph async we want to block frame invalidation temporarily since we
will do so later in the method manually if it's necessary.

I also noticed we have a few uses of `friend class` here or there that
need to eventually be cleaned up. I've made a note of it, and I will
handle it once Krita 5 is officially out.

BUG:441587
(cherry picked from commit b9eed7ba0ed597aedb488ddc37aaa31a2a659d0e)