Change autosave to be every 7 minutes
The default autosave option was set to 15 minutes because once upon
the time autosave was in the main thread, so it prevented users from
working for a short while. This commit changes the autosave interval
to 7 minutes because now the autosave happens in the background thread
so it doesn't interrupt the user, and this way it's less chance of
losing work.