Set Crop Tool to have Grow checked by default
Crop Tool's old behaviour was to only crop the canvas, never enlarge it.
Some time ago the Grow behaviour was added, but since it was a new
feature, it was disabled by default.
The Grow behaviour doesn't really limit the usage of Crop Tool since
the rectangle still snaps to the image boundaries (at least by default),
and it does increase the usage. New users expect it to use Grow
by default too.