Change Lod in Move Tool to 'true' by default
Before this commit, the Move Tool in Krita 5 would use Instant Preview
by default. Because the Instant Preview doesn't adapt to new
zoom levels, it caused a visible "regression" that a Move Tool
was no longer WYSIWYG when used, if the user changed the zoom level,
it would no longer show clear pixels.
Since Move Tool is not as performance heavy as possibly other tools
that use Instant Preview, it's better to have it turned off by default
to not limit the usefulness of the tool.