Add a (hidden) option to invalidate texture buffers right after the usage
When the function is enabled, the buffer is marked as invalidated
right after Krita requested its uploading to the GPU texture. It
should potentially increase the speed of texture uploading and reduce
GPU memory footprint Krita uses.

To enable the option, add the following line into `kritarc`:

useBufferInvalidation=true

You need to also pay attention to the following lines in the system
information log:

supportsBufferInvalidation: true
useBufferInvalidation (config option): true

The first line tells if the feature is supported bu the GPU, the second
tells if the feature is enabled in Krita's config file.

CC:kimageshop@kde.org