Fix assistants preview vs drawn assistant
Before this commit, some (especially new) assistants were confused
what is the preview (which can be disabled with View -> Show
Assistants Preview) and what is the assistant itself
(which can be disabled with View -> Show Painting Assistants).
This commit fixes it for three assistants:
- Parallel Ruler
- Vanishing Point
- Two Point Assistant

Previews: all moving parts (dependant on the mouse position)
Assistants: all static parts (independant on the mouse position)

Note that Two Point and Vanishing Point assistants don't use
much of drawCache and most of the assistant is drawn in the
drawAssistant() function. This most probably means lower
performance and it should probably be fixed by moving
drawing the static parts of the assistants into the drawCache()
function.

CCBUG:433097
CCBUG:442102