Fix (probably) improper found Python Lib without breaking dependencies
NOTE: Never set PYTHONPATH on the global scope of the build script as it
breaks anything python related, even Python compilation. It worked before
because it was a dirty configuration compile.

Cmake does not look for frameworks in buildroot/i/lib which causes our
Python to be missed, however we can't simply add the dir globally as
QtZlib.Framework is used instead of our zlib breaking the build. A
simple solution would be to delete QtZlib.Framework which we don't use
but our dependencies might. Another solution, used here is to add our framework
directory only for PythonLib search.