Fix Show Painting Assistants to update correctly
Before this commit, if you opened two files with assistants,
and in one turned the View -> Show Painting Assistants off,
and then switch to the other file and to that same file again,
the View -> Show Painting Assistant would be "on" even though
the painting assistants wouldn't be drawn on the canvas
(only preview would be visible). And then switching it off
would actually turn the drawing of the assistants on.

Another way to reproduce this issue would be to open a new document,
create a new assistant, turn off Show Painting Assistant,
then create a new assistant. Show Painting Assistant now looks "on"
but the drawing of the assistants is still actually off (which is
correct and it's just the checkbox that should stay "off").

This commit fixes this behaviour and now the checkbox in the View menu
reflects correctly the current state of drawing the assistant.

BUG:399419
BUG:417792