Move Perspective assistant drawing to drawAssistant from drawCache
Before this commit, the grid of the Perspective Assistant was drawn
in the drawCache, which caused performance issues when using
the perspective assistant and zooming in a lot.
(test method: zoom in a lot, so it's around 1600%, and then zoom in
and out very quickly using your mouse wheel. It's instant with this
patch, but slow without the patch. Note that if the Configure Krita ->
Display -> Canvas Acceleration is off, it's still super slow).

This commit does two things:
- crops the drawn lines to the viewport
- moves the whole drawing action to drawAssistant
Note: just cropping the drawn lines, but still using the cache,
didn't improve the assistants performance that much.
There must be some inherent issue with this caching method.

BUG:411352