Fix artifacts on the canvas when the display scaling is set to 110%
The problem is that we do "cropping" of the painted area twice during
one frame. Firstly, we enable stencil test to avoid too much pixels
to be rendered. And secondly, we limit the number of tiles that
are rendered on screen. And these two limit must agree in how they
do rounding.

CCBUG:442796