Fix crash when trying to export storyboard without scenes.
We also prevent the export window from opening at all when there's no
storyboard content available now. Lastly, in the odd chance the window
does open, we make sure it won't crash due to divide by zero.

We will disable the export button to reflect that this isn't
possible when there's nothing to export in a later commit.

BUG:442701