Fix forgetting about the cell spacing on opening the document
Before this commit, Krita loading a file with an isometric grid
wouldn't read what was the cell spacing for that grid and just use
the default value.
This commit makes it possible for Krita to read and use that value
and now the grids saved in Krita can be opened and look the same way
as before saving.