Fix wrong sorting of the tags after renaming a tag
Before this commit, the following order of action:
- assign a brush to a new tag called "Lily"
- select tag Lily
- rename the tag to "Zuzanna" and check where the tag is on the list,
- rename the tag to "Anna" and check where the tag is on the list,
would show that the tag list is not resorted after the rename.

HACK:
This commit ensures the list will be resorted because it first
asks the model to sort by other field.

NOTE: in KisTagModel in the current state (93277c99c8) there is a line
"dataChanged(idx, idx, {Qt::EditRole})" which might seem as if
it was the culprit of the model not recognizing that yes, it needs to
resort again (because EditRole != DisplayRole != UserRole + Name !=
column Name).
However that line doesn't change anything. The only thing that I've
found that works is resorting by other field.