Fix gradients icons on creating a bundle
Before this commit, creating a bundle with gradients
was difficult because the gradients weren't visible.
That's because the gradient thumbnail is only 1px tall,
while it is a few hundreds pixels long.
This commit ensures the gradients thumbnails are scaled
while ignoring the aspect ratio,
which avoids the issue (the thumbnails are now correctly scaled
vertically as well).