Fix gamut masks icons after switching modes
Before this commit, if you switched the view mode in the gamut masks
docker from Thumbnails to Details and back to Thumbnails,
the Thumbnails would be very stretched out.
This commit ensures the thumbnails are shown the same way they were
shown before the other mode was selected, which is just as icons
in the icon view, not list view.