Fix max size of the masking brush in the brush editor widget
We set the limit to the masking brush to 3x value of the max size
of the nurmal brush. It is needed to make sure that masking brush size
scales correctly when the brush is resized (they are linked
proportionally).

The limit was present here for ages, but the widget had incorrect
range set.