Fix md5sum in db to always represent the current resouce file
Before this commit, the md5sum in the database would always be
the same as the md5sum of the resource file of the first version
of the resource.
This commit ensures that the md5sum is updated in the database
to the latest version when a new version is created.