Add import, rename and add functions with user input
Before this commit, there was no central place where the code
for importing, adding or renaming resources with user input
was located. Every widget had to use KisResourceModel
and handle the user input (asking the user for confirmation,
showing the error messages etc.) on their own. This led to
differences between handling different resources
and problems making sure the logic is the same everywhere.
This commit adds the functions in one central place.