Don't download translations more than once a day
Before this commit, if the developer had translations on,
those translations would be downloaded via svn on every compilation,
adding unnecessary 40-50 seconds (on my setup) even to the
recompilation (compilation with no changes in the code, just cmake
going through all the files).
This commit adds creating a file with the last_update_time
so next time the script to download the translations is called,
it can check if it downloaded the translations that day and if yes,
just omit downloading at all.