Avoid leaking "Create new document" dialogs
If the dialog was just closed with the button in the window title bar,
it never triggered any signal that would delete it.
And generally, having KisPart delete it is a strange design, it cannot
really know if it was allocated on the heap in the first place, nor should
it be responsible for closing the dialog.

Instead, make the dialog buttons call QDialog::accept() or
QDialog::reject() accordingly, and delete the dialog where it was allocated.

(cherry picked from commit e5f10d40b1c5a926e39dff3655b573d00ffef776)