Fix consistency of 'dependent_resources_filenames' metadata
Now the update of this metadata is requested by KisResourceLocator
every time a new version of the resource is added to the database.

As noted in the commets, this check works only for legacy presets,
which don't include all their dependent resources. For the modern
presets we don't check if the dependent resources has been loaded
successfully or not. We just hope that they have been embeded
correctly.

WARNING: the previous code sometimes generated incorrect
"dependent_resources_filenames" information. If your resources became
marked as "broken", just activate them to update the metadata.

CC:kimageshop@kde.org