Remove KisResourceModel as argument in KisResourceUserOperations
Before this commit, KisResourceUserOperations used KisResourceModel
pointer in a parameter in a function to access the model.
This commit removes that parameter and always creates a new model.
This ensures that the model does always show all the resources
and doesn't have any filtering applied.
Also note that lots of the models had to be created immediately
before the function was called anyway.