Fix linking color space of Alpha color spaces
Before the patch, the alpha color spaces could be interpreted in
multiple different ways, depending on the conversion destination.
When it was converted into rgb8 or graya8, it was interpreted as
srgb-gamma corrected graya space. But when the conversion was
requested for rgb16 or graya16 color space (e.g. in color smudge
tool), the conversion happened through the LabA color space, which
is linear. It caused rendering issues, when painting with color
smudge brushes on transparency masks (or any alpha8-based layers).

This patch fixes the issues and makes "Gray-D50-elle-V2-srgbtrc" to
be the only linking color space for alpha colorspaces. Which basically
means that all the flavours of alpha color space are now basically
flattened versions of srgb-gamma-corrected graya color space.

BUG:443422
CC:kimageshop@kde.org