Fix custom_pattern having multiple question boxes
Before this commit, custom_pattern widget had multiple
question boxes because the code was not accustomed to the fact
that WithUserInput functions handle it on its own.
This commit ensures that the dialog boxes happens always when needed
but not more often.

There is a problem with using QFileDialog that it checks for overwrite
itself, but if we add the suffix later, then it doesn't ask the user
since it doesn't know we'll add the suffix. So the dialog later
had to check if any change for the filename was used,
otherwise the dialog asking the user if they want to overwrite
would appear twice.

This commit also fixes lack of suffix in case the user selected
"All supported formats" option in the combobox in the QFileDialog
(it was caused by using 'QFileInfo(filename).fileName()' since the
`filename` doesn't get that additional suffix.

It also makes sure that the filename doesn't get two dots
`filename..suffix` in the line that changes from fi.fileName() to
fi.baseName() when adding the suffix (baseName() removes the suffix at
the end that would be added when using "All supported formats"
option in the combobox).