Fix crash during animation export by simplifying RAII of ffmpeg subprocess.
We were manually cleaning up FFMPEG processes in some parts of the code
instead of relying on RAII, which caused some circumstances where the
process was freed while QProcess was invoking `waitForProcess`, which
seems to crash only on Windows systems. Qt must normally handle this
more gracefully on Linux, but can't do so on Windows.

We're now avoiding manual cleanup by making more use of QScopedPointer
and making our end of the code more clean. Now the process should simply
be cleaned up when FFMpegWrapper goes out of scope or is deleted,
which should avoid the pointer going null before it is supposed to.

BUG:442578