Storyboard Export: Added default storyboard export template.
This SVG file can be used when exporting a group of storyboards in
Krita's Storyboard Docker. Select "SVG Template" as the layout type and
then select `storyboard_templates/storyboard_extemp_default.svg`.

This particular template supports 2 comment tracks:
"Comment", on the right side of the board images, and..
"Sound", on the left side.
In order to make use of these comment spaces, your comment tracks within
Krita's Storyboard Docker must be named accordingly. ("Comment" and
"Sound", respectively.)

This will be added to the docs ASAP. :)