Refine logic in kis_custom_pattern
Before this commit, there was no early returning if the user
pressed Cancel. This commit ensures the function returns early
so there cannot be any trouble down the line.

Before this commit, the dialog wouldn't notice that the user wanted
to overwrite if they clicked OK on the Overwrite dialog.
This commit ensures the logic around that is correct.