Allow overwriteWithUserInput to accept "wild" resource pointers
Before this commit, overwriteResourceWithUserInput only accepted
pointers that were coming from the cache.
This commit ensures that it can handle a situation when the pointer
doesn't come from the cache (for example was created by a widget).
It retrieves the pointer from the cache, then moves the data from
the pointer it got as a parameter to the cached pointer.
Then it saves it as a new version of the resource.