Split the document storage in to storages
One for embedded resources, one for linked resources. Embedded
resources are resources that are saved and managed by parts of the
kra file, like fill layers. Linked resources are resources the
user has saved to the document, like palettes.

In the future, our resource saving widgets should allow for the
selection of document storages.