KisAslLayerStyleSerializer: fix layer-local gradients load
In 2d426ac0b57ad8595a01b0c82d18633e4ec6060b, @dkazakov refactored the
load so that layer styles would insert their resources on global storage
instead of keeping them in their local stashes. However, this did not
take in account that layers can also have locally defined e.g.
gradients. These will not be found if the resource interface is
overriden because they are resident on the root layer's
localStrokeStorage.

Upon a complete review of the code, the true reason for this failure
lied not only in the (in)validity of the layer style, which I fixed in a
previous commit, but also on the fact that gradients and patterns are
uploaded *before* they are read (!!), in the embedded pattern
registration loop.

This commit fixes that by moving the two upload loops to inside the
layer style deserialization, and making sure they upload only missing
resources (to avoid accidentally quadratic behaviour).

BUG: 441605
CCMAIL: dimula73@gmail.com
CCMAIL: tamtamy.tymona@gmail.com