Ensure that the preset will have -1 id when saving as a new one
Before this commit, the preset was simply cloned, so it still had
the id of the preset it was copied from. It confused the resource system
into thinking it's the same preset. This commit makes sure the id
is reset to an invalid value so it can be added to the resource system
seamlessly.