fix huge layer docker minimum width
fix huge layer docker minimum width

For some reasons setting minimumSize properties to cmbComposite made its
minimum width as long as the longest composite op name (and some can be veeery
long, especially translated). Fitting to the box to text size is not really
needed as there are tooltips showing the full name, and the extra docker size was
a big waste of painting area.