Fix broken adding a workspace from the main menu
Before this commit, there was no way to define a name for the new
workspace in Window -> Workspaces -> New Workspace..., so Krita
would always try to call it "" (empty string). This of course
always failed.
This commit fixes that by readding previously removed (by mistake)
QInputDialog asking for the name of the resource.

BUG:444112