Fix session saving into a wrong place when closing Krita
Before this commit, if the user enabled "save session on closing Krita"
in Configure Krita -> General -> Misc, Krita would save it
in a working directory of the executable (usually the home directory,
but can be whatever directory Krita was started from).
This commit fixes it by ensuring it uses the new resource system
properly.

BUG:443652