Welcome Page: Slightly reduced support text font size.
A previous commit scaled the support message up +130%,
but I felt that it was a bit too large compared to the other
important text on the page, so I've change it to be +110% instead.

We can discuss if we think this is too small or if there's a better way
to present this message to users.


(cherry picked from commit e2a04dd8231254543de8c5ff151aba91785c7ebe)